Move flexpolyline tests to test sources
